+++ src/api/libipc/inc/ipc_settings.h
