+++ src/api/libipc/src/inc/ipcd.h
@@ -18,7 +18,7 @@ extern "C"
 
 // Project headers
 #include "client_t.h"
-#include "settings.h"
+#include "ipc_settings.h"
 #include "immut.h"
 #include "ipc_packet.h"
